+++ tests/testthat/test-messages.R
@@ -6,7 +6,7 @@ mockery::stub(
 )
 
 cli::test_that_cli("messages work", {
-  skip_if_not_installed("rlang", "1.0.0")
+  # skip_if_not_installed("rlang", "1.0.0")
   withr::local_envvar(GITHUB_ACTIONS = "false")
 
   withr::with_options(list(octolog.debug = TRUE), {