+++ NEWS.md
@@ -1,6 +1,9 @@
 # octolog (development version)
 
-# octolog 0.1.1
+* octolog is now compatible with rlang >= 0.4.10
+* openssl is now optional as it is only used in `octo_stop_commands()`
+  
+* # octolog 0.1.1
 
 * add CITATION